% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageToData.R
\name{ImageToData}
\alias{ImageToData}
\title{Convert Image to Data}
\usage{
ImageToData(
  url,
  split = " ",
  filter = c("", "="),
  lang = "eng",
  first.colnames = TRUE,
  write.clip = TRUE,
  as.DT = FALSE
)
}
\arguments{
\item{url}{A character string containing the url of the image to be converted}

\item{split}{A character string containing a perl compatible regular
expression to split cells}

\item{filter}{A character vector of strings which should be filtered out of
the final result}

\item{lang}{A chracter vector of the language to be used by tesseract.}

\item{first.colnames}{A logical scalar. Should the first row become column
names?}

\item{write.clip}{A logical scalar. Should the resulting data be written to
the clipboard for further editing?}

\item{as.DT}{A logical scalar. Should the data be returned as a data.table?
If FALSE, a data.frame will be returned.}
}
\value{
If \code{as.DT = TRUE}, a data.table will be returned. Otherwise a
  data.frame will be returned.
}
\description{
\code{ImageToData} returns a \code{data.frame} (by default) or
\code{data.table} from an image url using optical character recognition.
}
\details{
This is a helper function intended to streamline the process of reading
sample data posted by users on Stack Overflow. Slight errors are common, and
thus by default the data is written to the clipboard for further processing
in the user's preferred spreadsheet software.
}
\examples{
ImageToData("https://i.stack.imgur.com/yBcfz.jpg")


}
