% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{ReadData}
\alias{ReadData}
\title{Read Data from the Clipboard}
\usage{
ReadData(first.colnames = TRUE, as.DT = FALSE, ...)
}
\arguments{
\item{first.colnames}{A logical scalar. Should the first row become column
names?}

\item{as.DT}{A logical scalar. Should the data be returned as a data.table?
If FALSE, a data.frame will be returned.}

\item{...}{Additional paramaters to be passed to \code{read.table}.}

\item{write.clip}{A logical scalar. Should the resulting data be written to
the clipboard for further editing?}
}
\value{
If \code{as.DT = TRUE}, a data.table will be returned. Otherwise a
  data.frame will be returned.
}
\description{
\code{ReadData} reads tab seperated data from the clipboard and returns a
\code{data.frame} (default) or \code{data.table}.
}
\details{
This is a helper function intended to streamline the process of reading
sample data posted by users on Stack Overflow. Slight errors are in optical
character recognition are common. Thus, minor editing in a spreadsheet
application may be necessary. This function reads tab seperated data in the
cliboard into a \code{data.frame} (default) or \code{data.table}.
.
}
